requirements.txt
APScheduler==3.10.0
beautifulsoup4==4.11.2
coloredlogs==15.0.1
geopy==2.3.0
gntp==1.0.3
mattermostdriver==7.3.2
O365==2.0.26
paho_mqtt==1.6.1
pyfirmata==1.1.0
PyMySQL==1.0.2
python-dotenv==0.21.1
requests==2.28.2
slate3k==0.5.3
spotipy==2.22.1
wikipedia==1.4.0
wolframalpha==5.0.0
phue==1.1.0
python-aiml==0.9.3
openai==0.26.5
